+++ autometrics/build.rs
@@ -11,6 +11,8 @@ pub fn main() {
     println!("cargo:warning=The `prometheus` feature is deprecated and will be removed in the next version. Please use `prometheus-0_13` instead.");
     #[cfg(feature = "prometheus-client")]
     println!("cargo:warning=The `prometheus-client` feature is deprecated and will be removed in the next version. Please use `prometheus-client-0_21` instead.");
+    #[cfg(feature = "exemplars-tracing-opentelemetry")]
+    println!("cargo:warning=The `exemplars-tracing-opentelemetry` feature is deprecated and will be removed in the next version. Please use `exemplars-tracing-opentelemetry-0_20` instead.");
 
     cfg_aliases! {
       // Backends
@@ -30,7 +32,7 @@ pub fn main() {
       // Exemplars
       exemplars: { any(exemplars_tracing, exemplars_tracing_opentelemetry) },
       exemplars_tracing: { feature = "exemplars-tracing" },
-      exemplars_tracing_opentelemetry: { feature = "exemplars-tracing-opentelemetry" },
+      exemplars_tracing_opentelemetry: { any(feature = "exemplars-tracing-opentelemetry-0_20", feature = "exemplars-tracing-opentelemetry") },
 
       // Custom objectives
       custom_objective_percentile: { feature = "custom-objective-percentile" },